refactor: mock service worker setupServer into testServer.js
